+++ examples/reuse_scope.rs
@@ -1,4 +1,4 @@
-use rhai::{Engine, EvalAltResult, Scope};
+use rhai::{Engine, EvalAltResult, Scope, INT};
 
 fn main() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
@@ -6,7 +6,7 @@ fn main() -> Result<(), EvalAltResult> {
 
     engine.eval_with_scope::<()>(&mut scope, "let x = 4 + 5")?;
 
-    let result = engine.eval_with_scope::<i64>(&mut scope, "x")?;
+    let result = engine.eval_with_scope::<INT>(&mut scope, "x")?;
 
     println!("result: {}", result);
